+++ bsp/rp2040/unique.c
@@ -11,7 +11,7 @@ uint8_t get_unique_id_u8(uint8_t *desc_str) {
 
 	pico_get_unique_board_id(&uid);
 
-	for (int byte = 0; byte < TU_ARRAY_SIZE(uid.id); byte++) {
+	for (size_t byte = 0; byte < TU_ARRAY_SIZE(uid.id); byte++) {
 		uint8_t tmp = uid.id[byte];
 		for (int digit = 0; digit < 2; digit++) {
 			desc_str[chr_count++] = nyb2hex(tmp & 0xf);
@@ -28,7 +28,7 @@ uint8_t get_unique_id_u16(uint16_t *desc_str) {
 
 	pico_get_unique_board_id(&uid);
 
-	for (int byte = 0; byte < TU_ARRAY_SIZE(uid.id); byte++) {
+	for (size_t byte = 0; byte < TU_ARRAY_SIZE(uid.id); byte++) {
 		uint8_t tmp = uid.id[byte];
 		for (int digit = 0; digit < 2; digit++) {
 			desc_str[chr_count++] = nyb2hex(tmp & 0xf);
