+++ bsp/stm32f072disco/unique.c
@@ -8,7 +8,7 @@ uint8_t get_unique_id_u8(uint8_t *desc_str) {
 	uint32_t tmp = 0;
 	uint8_t chr_count = 0;
 
-	for (int digit = 0; digit < 24; digit++) {
+	for (size_t digit = 0; digit < 24; digit++) {
 		if (0 == (digit & 7)) tmp = *idpnt++;
 		desc_str[chr_count++] = nyb2hex(tmp & 0xf);
 		tmp >>= 4;
@@ -22,7 +22,7 @@ uint8_t get_unique_id_u16(uint16_t *desc_str) {
 	uint32_t tmp = 0;
 	uint8_t chr_count = 0;
 
-	for (int digit = 0; digit < 24; digit++) {
+	for (size_t digit = 0; digit < 24; digit++) {
 		if (0 == (digit & 7)) tmp = *idpnt++;
 		desc_str[chr_count++] = nyb2hex(tmp & 0xf);
 		tmp >>= 4;
